Introduce home (main) activity, remove more activity (rel. to #11904)
