--- main/res/layout/more_activity.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ListView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/more_actions"
-    android:layout_width="fill_parent"
-    android:layout_height="wrap_content"
-    android:divider="@null"
-    android:orientation="vertical"
-    tools:listitem="@layout/page_more_item" />
