--- main/res/layout/page_more_item.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="fill_parent"
-    android:layout_height="wrap_content"
-    android:background="?android:attr/selectableItemBackground"
-    android:paddingVertical="7dip"
-    android:paddingHorizontal="5dip"
-    tools:context=".downloader.DownloadSelectorActivity">
-
-    <ImageView
-        android:id="@+id/item_icon"
-        style="@style/button_icon"
-        android:layout_width="36dp"
-        android:layout_height="36dp"
-        android:gravity="left"
-        android:clickable="false"
-        android:layout_centerVertical="true"
-        android:layout_alignParentLeft="true" />
-
-    <TextView
-        android:id="@+id/item_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerVertical="true"
-        android:layout_toRightOf="@id/item_icon"
-        android:layout_toLeftOf="@id/navigate_button"
-        android:ellipsize="end"
-        android:gravity="left"
-        android:paddingLeft="10dip"
-        android:textColor="@color/colorText"
-        android:textIsSelectable="false"
-        android:textSize="@dimen/textSize_headingPrimary"
-        tools:text="name" />
-
-    <ImageButton
-        android:id="@+id/navigate_button"
-        style="@style/button_icon"
-        android:layout_width="36dp"
-        android:layout_height="36dp"
-        android:clickable="false"
-        android:layout_centerVertical="true"
-        android:layout_alignParentRight="true"
-        android:src="@drawable/ic_navigate_next"/>
-</RelativeLayout>
