+++ main/res/menu/bottom_navigation_menu.xml
@@ -2,6 +2,11 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <!-- WARNING: Bottom navigation menu can hold only five items max! -->
     <item
+        android:id="@+id/page_home"
+        android:enabled="true"
+        android:icon="@drawable/ic_menu_home"
+        android:title="@string/home_button" />
+    <item
         android:id="@+id/page_map"
         android:enabled="true"
         android:icon="@drawable/ic_menu_mapmode"
@@ -21,9 +26,4 @@
         android:enabled="true"
         android:icon="@drawable/ic_menu_nearby"
         android:title="@string/caches_nearby_button" />
-    <item
-        android:id="@+id/page_more"
-        android:enabled="true"
-        android:icon="@drawable/ic_menu_more_horizontal"
-        android:title="@string/more_button" />
 </menu>
