--- main/res/menu/bottom_navigation_more.xml
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:id="@+id/menu_dashboard"
-        android:icon="@drawable/ic_menu_dashboard_stats"
-        android:title="@string/menu_dashboard">
-    </item>    <item
-    android:id="@+id/menu_history"
-    android:icon="@drawable/ic_menu_recent_history"
-    android:title="@string/menu_history_longstring">
-    </item>
-    <item
-        android:id="@+id/menu_goto"
-        android:icon="@drawable/ic_menu_goto"
-        android:title="@string/any_button">
-    </item>
-    <item
-        android:id="@+id/menu_pocket_queries"
-        android:icon="@drawable/ic_menu_pocket_query"
-        android:title="@string/menu_lists_pocket_queries"
-        android:visible="false">
-    </item>
-    <item
-        android:id="@+id/menu_bookmarklists"
-        android:icon="@drawable/ic_menu_bookmarks"
-        android:title="@string/menu_lists_bookmarklists"
-        android:visible="false">
-    </item>
-    <item
-        android:id="@+id/menu_about"
-        android:icon="@drawable/ic_menu_info_details"
-        android:title="@string/menu_about">
-    </item>
-    <item
-        android:id="@+id/menu_wizard"
-        android:icon="@drawable/ic_menu_preferences"
-        android:title="@string/wizard">
-    </item>
-    <item
-        android:id="@+id/menu_settings"
-        android:icon="@drawable/ic_menu_preferences"
-        android:title="@string/menu_settings">
-    </item>
-    <item
-        android:id="@+id/menu_backup"
-        android:icon="@drawable/ic_menu_rotate"
-        android:title="@string/menu_backup">
-    </item>
-    <item
-        android:title="@string/menu_update"
-        android:icon="@drawable/ic_menu_refresh">
-        <menu>
-            <item
-                android:id="@+id/menu_update_routingdata"
-                android:title="@string/menu_update_routingdata">
-            </item>
-            <item
-                android:id="@+id/menu_update_mapdata"
-                android:title="@string/menu_update_mapdata">
-            </item>
-        </menu>
-    </item>
-    <item
-        android:id="@+id/menu_helpers"
-        android:icon="@drawable/ic_external_link"
-        android:title="@string/menu_helpers">
-    </item>
-    <item
-        android:id="@+id/menu_scan"
-        android:icon="@drawable/ic_menu_barcode"
-        android:title="@string/menu_scan_geo">
-    </item>
-    <item
-        android:id="@+id/menu_report_problem"
-        android:icon="@drawable/ic_menu_bug_report"
-        android:title="@string/debug_user_error_report_title">
-    </item>
-</menu>
