+++ main/res/values/strings.xml
@@ -353,12 +353,12 @@
     <string name="menu_invert_order">Invert order</string>
 
     <!-- main screen -->
+    <string name="home_button">Home</string>
     <string name="live_map_button">Live map</string>
     <string name="caches_nearby_button">Nearby</string>
     <string name="advanced_search_button">Search</string>
     <string name="stored_caches_button">Stored</string>
     <string name="any_button">Go to</string>
-    <string name="more_button">More</string>
     <string name="unknown_scan">Didn\'t find any geocodes in the scan result.</string>
 
     <!-- caches -->
